+++ src/Text/zh-TW.lua
@@ -29,5 +29,23 @@ mod.AddLocale("zh-TW", {
 	ResourceMenuTitle = "資源選單",
 	ResourceMenuDescription = "創造任何數量的資源。",
 	ResourceMenuSpawnButton = "創造資源",
-	ResourceMenuEmpty = "空"
+	ResourceMenuEmpty = "空",
+
+	BossSelectorTitle = "區域守衛選擇器",
+	BossSelectorDescription = "讓您直接去打一位區域守衛，用您的存檔裝備。",
+	BossSelectorNoSavedState = "沒有存檔裝備! 請去造一個!",
+
+	KillPlayerTitle = "殺玩家",
+	KillPlayerDescription = "殺您而遣送您到三岔路口。",
+
+	SaveStateTitle = "存入裝備",
+	SaveStateDescription = "存入您現在的裝備，為了一後加載。 必要為了用區域守衛選擇器。 存入您全部裝備。",
+	SaveStateSaved = "裝備存入了!",
+
+	LoadStateTitle = "載入裝備",
+	LoadStateDescription = "載入您的存檔裝備。 必要先存入一個裝備。",
+	SaveStateLoaded = "裝備加載了!",
+
+	ConsumableSelectorTitle = "耗材選擇器",
+	ConsumableSelectorDescription = "自己給您如何耗材。"
 })